+++ flutter_packages/file_download/lib/src/download/download_image.dart
@@ -1,7 +1,9 @@
 import 'dart:typed_data';
 
-export 'download_image_io.dart'
-    if (dart.library.js_interop) 'download_image_web.dart'; // ignore: uri_does_not_exist
+import 'package:tekaly_file_download_web/file_download.dart'
+    show DownloadFileInfo;
+
+import 'download_file.dart';
 
 class DownloadImageInfo {
   /// Must have the proper extension
@@ -22,3 +24,14 @@ class DownloadImageInfo {
   String toString() =>
       'DownloadImageInfo($filename, ${data.length} bytes, $mimeType)';
 }
+
+/// [DownloadImageInfo] as a [DownloadFileInfo].
+extension DownloadImageInfoExtension on DownloadImageInfo {
+  /// The matching file info, as used by `downloadFile`.
+  DownloadFileInfo toDownloadFileInfo() =>
+      DownloadFileInfo(filename: filename, data: data, mimeType: mimeType);
+}
+
+/// Download an image, see `downloadFile` for the generic helper.
+Future<void> downloadImage(DownloadImageInfo imageInfo) =>
+    downloadFile(imageInfo.toDownloadFileInfo());